Rename websocket dir to websocket_server
