+++ src/components/transport_manager/CMakeLists.txt
@@ -108,13 +108,13 @@ endif()
 
 if(NOT BUILD_WEBSOCKET_SERVER_SUPPORT)
   list(APPEND EXCLUDE_PATHS
-    ${CMAKE_CURRENT_SOURCE_DIR}/include/transport_manager/websocket
-    ${CMAKE_CURRENT_SOURCE_DIR}/src/websocket
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/transport_manager/websocket_server
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/websocket_server
   )
 elseif(NOT ENABLE_SECURITY)
   list(APPEND EXCLUDE_PATHS
-    ${CMAKE_CURRENT_SOURCE_DIR}/include/transport_manager/websocket/websocket_secure_session.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/src/websocket/websocket_secure_session.cc
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/transport_manager/websocket_server/websocket_secure_session.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/websocket_server/websocket_secure_session.cc
   )
 endif()
