+++ src/components/transport_manager/test/websocket_connection_test.cc
@@ -36,8 +36,8 @@
 #include <string>
 
 #include "resumption/last_state_impl.h"
-#include "transport_manager/websocket/websocket_connection.h"
-#include "transport_manager/websocket/websocket_session.h"
+#include "transport_manager/websocket_server/websocket_connection.h"
+#include "transport_manager/websocket_server/websocket_session.h"
 
 #include "transport_manager/mock_transport_manager_settings.h"
 #include "transport_manager/transport_adapter/mock_transport_adapter_controller.h"
